+++ src/demux/demuxer-worker.js
@@ -39,7 +39,7 @@ var DemuxerWorker = function (self) {
         forwardMessage('init',null);
         break;
       case 'demux':
-        self.demuxer.push(new Uint8Array(data.data), data.audioCodec, data.videoCodec, data.timeOffset, data.cc, data.level, data.sn, data.duration,data.accurateTimeOffset,data.defaultInitPTS);
+        self.demuxer.push(data.data, data.audioCodec, data.videoCodec, data.timeOffset, data.cc, data.level, data.sn, data.duration,data.decryptdata,data.accurateTimeOffset,data.defaultInitPTS);
         break;
       default:
         break;
